+++ examples/msg_between.rs
@@ -1,10 +1,11 @@
 use kdam::prelude::*;
+use kdam::term::Writer;
 
 fn main() {
     let mut pb1 = tqdm!(total = 100, position = 0);
     let mut pb2 = tqdm!(total = 100, position = 4);
 
-    kdam::eprint_at!(2, "any text can be placed {} bars", "between");
+    Writer::Stderr.print_at(2, "any text can be placed between bars");
 
     for _ in 0..100 {
         std::thread::sleep(std::time::Duration::from_secs_f32(0.02));